Add Coordinate systems and reference frames to WebVR
Currently, the WebVR standard assumes that at any given time there is
a single world space frame of reference (which can be reset by resetPose).
While this is a simple model that corresponds to how developers are used to
thinking about things, it does not map 1:1 to how inside-out-trackers see
the world. Specifically, as you move away from the origin an inside-out-tracker
has less information available to accurately locate where that original
"world origin" is in relation to the HMD current position. For that reason
placement of objects can start showing precision issues and drift.

To solve this, we would need to reason about multiple frames of reference
in some way, so that apps can more accurately render their experiences
as users move away from the original origin. This pull request represents
an path to explicitly representing to developers the "squishy" nature of
the tracking technologies.   In this proposal there is no special "blessed" world
space frame of reference. The user can create any number of frames of
reference of various kinds, and will explicitly decide which one they're using for
the experience they are trying to build.  They will then supply this frame of
reference whenever querying for transform data (such as getFrameData).  In the near
future we expect that this will then be extended to include other types such as
anchors to specific locations in the world and surface reconstruction meshes.

One benefit of this system is that it maps very directly to inside-out-tracking
algorithms, which reduces the risk that we over-simplify the mental model
for developers. It also means we can support truly large scale applications
where the user would move far from their original position (e.g. Pokemon Go in MR).
In addition, it means that things like the stage can be naturally expressed
as yet another frame of reference.