VRPresentationFrame, promise-based render loop, WEBGL_multiview
First up, @toji had the good idea to divide the data that changes each presentation frame from the data that changes each presentation frame and is also dependent on the coordinate system.  Thanks!

Second, in an attempt to scope down the changes to the smallest tightly-coupled cohesive set necessary to merge, I've reverted back to the Promise pattern rather than the RAF pattern.  Once we've gotten closure on the rest of this design, I'll open a separate issue to discuss this specific topic since I'm still on the fence.

Third, after poking at it a bit more, there may actually be a path to using the currently proposed multiview extension (https://www.khronos.org/registry/webgl/extensions/proposals/WEBGL_multiview/) rather than a custom side-by-side one. That said, there are a few things to note:
- The current multiview extension is designed such that developers attach textures to a framebuffer to create a multiview-aware framebuffer. In WebVR 2.0, the UA should be providing the optimal framebuffer for the hardware they are targeting rather the other way around. As a result, we can achieve the desired outcome by adding an optional "multiview" attribute to the VRWebGLLayerInit which will construct the appropriate framebuffer for multiview rendering. This framebuffer is then passed into `bindFramebuffer()` as usual and it behaves as if it was constructed using the multiview extension.
- Because the array vs. side-by-side decision is abstracted from the web developer, we open up the option for the UA to pick the optimal texture format.  Yay!
- The main pain point with this design is that there is no easy way to restrict rendering to a subset of the views (say, just the right eye) because the framebuffer isn't populated by the web developer.  In order for this design to be feasible, that will need to be addressed somehow. :-/ @rafaelcintron has started a discussion with Olli (who is proposing the WEBGL_multivew extension in question) and we'll see where it goes