Change VRWebGLLayer to take a context rather than a canvas
Given the recent changes in the render loop there no longer seems to be
a reason to track the canvas itself, so I propose here switching the
VRWebGLLayer to accept the rendering context instead.

Also did some research and found that, as @kbr warned, we can’t use
WebGLRenderingContextBase in our IDL directly. So instead we get to
play with more typedefs! *sigh*