Change animation loop from promise-based to callback-based
While talking with @NellWaliczek today she pointed out that Microsoft
was concerned about how to deal with VRPresentationFrame promises that
the app holds a reference to and calls `.then()` on well after the
frame has been processed. Upon reflection this is definitely a sticky
problem: Ideally we want any frame callbacks registered to simply pick
up the next frame’s data, but with a promise you could effectively hold
onto a stale frame for a long time.

It could be made to work, but the spec gymnastics required to ensure
sensible behavior would be significant and hard to parse, while the
existing `requestAnimationFrame` syntax is already well understood by
devs and gives us the desired behavior of never being able to fire a
frame callback for a stale frame. It will require a bit of re-hashing
the `requestAnimationFrame` syntax in the WebVR spec, but that’s
probably preferable to trying to push a Promise into a rAF-shaped hole.

Thanks for pointing out this issue, @NellWaliczek and @RafaelCintron!