Convert XRFrameOfReference to XRCoordinateSystem
This is a speculative change based on today’s WebXR call. It drops the
concept of a separate XRFrameOfReference object in favor of using
XRCoordinateSystem directly. This necessitated moving where bounds were
queried from, but I kind of like this better?

Naming feels a little awkward at this point, but it’s a start.