Single-canvas inline, drop XRPresentationContext
Fixes #636

This change adds a path for using a single WebGL canvas to drive inline
sessions, rather than needing both a WebGL context and an
XRPresentationContext before the inline session can function. The way to
do this is to create an `XRWebGLLayer` with the `useDefaultFramebuffer`
option set to `true` and set it as the session's `baseLayer`.

In addition to the above, this change removes the XRPresentationContext
entirely, and along with it the ability to do desktop mirroring (and
more advanced inline scenarios.) This is to simplify the core API, and
equivalent functionality will be reconsidered in the future because the
functionality that they represent is valuable.