Block mid-session consent requests
/fixes #751

Details that consent dialogs may not be displayed while an immersive
session is active, and explains that this is a short-term restriction.

I chose to go with terminating the session in this case rather than
suppressing the consent dialog altogether, since it provides very clear
feedback about the error case and doesn't require us to make up new
behavior for the permissions in question.