getViewport with an invalid view throws an error
/fixes #769 (sort of).

The previous logic to return `null` in situations where the passed view
was from a different session was not consistent with the error behavior
elsewhere. As a specific example, passing spaces from one session to an
XRFrame from a different session throws an `InvalidStateError`. As such,
throwing the same error here for the same type of validation makes the
spec more internally consistent.